requirements.txt
sqlalchemy
openai
python-dotenv